# server/docs/swagger.yaml
info:
  contact: {}
  description: |-
    客服系统api `
    客服后端：`
  title: 客服系统API文档`
  version: 1.0`
paths:
...
  /api/service/info:
    post:
      parameters:
      - description: 认证token
        in: header
        name: token
        required: true
        type: string
      responses: {}
      summary: 获取客服基本信息
      tags:
      - 客服信息
...
  /api/service/update:
    post:
      parameters:
      - description: 认证token
        in: header
        name: token
        required: true
        type: string
      - description: 客服名称
        in: query
        name: name
        required: true
        type: string
      - description: 客服头像
        in: query
        name: head
        type: string
      responses: {}
      summary: 修改客服头像和昵称
      tags:
      - 客服信息
...
swagger: "2.0"